fix: list index out of range while generating walking trajectories
This commit resolves the issue where the generate_walking_trajectories
sometimes throw `list index out of range` when the wandering occurrs.
This exception occurrs when the first step of a walking trajectory is
contained in the next target places. In particular, a start place of
wandering is randomly selected, so that the desk and desk_chair are
selected continuously and hit by a bug. It is resolved by regarding such
cases as not walking (as `centers = []`).